misc: get_acpi_data: let read_acpi_block() silently fail when SSDB/CLDB not found
Now that we iterate over all ACPI devices, let's lower the loglevel from
err to dbg to suppress the output when SSDB/CLDB is not found.

In this way, we can avoid spamming log. When needed, you can output this
log by adding "dyndbg" option to modprobe/insmod.

For example:

    $ sudo insmod <module_file> dyndbg

Signed-off-by: Tsuchiya Yuto <kitakar@gmail.com>